requirements.txt
scholarly>=1.7.0
pyyaml>=6.0
requests>=2.28.0
yt-dlp>=2024.1.0
pytest>=7.0.0
